feature: add prefix/postfix support to select fields
